pyproject.toml
[tool.poetry]
name = "modern-data-stack-showcase"
version = "1.0.0"
description = "A comprehensive demonstration of advanced data engineering patterns, sophisticated analytics architectures, and enterprise-grade platform engineering practices"
authors = ["Data Engineering Showcase <showcase@example.com>"]
readme = "README.md"
packages = [{include = "showcase", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"

# Core data processing
pandas = "^2.1.4"
numpy = "^1.26.2"
polars = "^0.20.3"
dask = {extras = ["complete"], version = "^2023.12.1"}

# Database connectivity
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"

# dbt and data transformation
dbt-core = "^1.8.7"
dbt-postgres = "^1.8.2"
dbt-bigquery = "^1.8.2"

# Data quality and testing
great-expectations = "^0.18.16"
pandera = "^0.17.2"
dbt-expectations = "^0.10.3"

# Machine learning
scikit-learn = "^1.3.2"
xgboost = "^2.0.3"
lightgbm = "^4.1.0"
optuna = "^3.5.0"

# Deep learning (optional)
tensorflow = {version = "^2.15.0", optional = true}
torch = {version = "^2.1.2", optional = true}

# MLOps and experiment tracking
mlflow = "^2.9.2"
wandb = "^0.16.1"

# Visualization
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
plotly = "^5.17.0"
altair = "^5.2.0"

# Statistical analysis
scipy = "^1.11.4"
statsmodels = "^0.14.1"

# Time series analysis
prophet = "^1.1.5"

# API and utilities
requests = "^2.31.0"
pydantic = "^2.5.2"
typer = "^0.9.0"
rich = "^13.7.0"

# Configuration management
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"

# Data generation
faker = "^20.1.0"

[tool.poetry.group.dev.dependencies]
# Development tools
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

# Testing
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

# Documentation
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.2"
mkdocs-mermaid2-plugin = "^1.1.1"

# Jupyter and notebooks
jupyter = "^1.0.0"
jupyterlab = "^4.0.9"
ipykernel = "^6.27.1"

# Linting and formatting
sqlfluff = "^3.0.7"
sqlfluff-templater-dbt = "^3.0.7"

# Performance profiling
py-spy = "^0.3.14"
memory-profiler = "^0.61.0"

[tool.poetry.group.powerbi.dependencies]
# Power BI and Azure integration (if available)
azure-identity = "^1.15.0"
azure-storage-blob = "^12.19.0"
requests-oauthlib = "^1.3.1"

[tool.poetry.extras]
deep-learning = ["tensorflow", "torch"]
all = ["tensorflow", "torch"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
] 